+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to AET will be documented in this file.
 ## Unreleased
 **List of changes that are finished but not yet released in any final version.**
 
+- [PR-328](https://github.com/Cognifide/aet/pull/328) Added suite's history
 - [PR-303](https://github.com/Cognifide/aet/pull/303) Added `exclude-elements` parameter to [ScreenCollector](https://github.com/Cognifide/aet/wiki/ScreenCollector)
 - [PR-327](https://github.com/Cognifide/aet/pull/327) Default web driver changed from Firefox to Chrome
 - [PR-294](https://github.com/Cognifide/aet/pull/294) Added support for full page screenshot in chrome